requirements.txt
fastapi==0.115.12
redis==6.2.0
prometheus-client==0.22.1
uvicorn[standard]==0.34.3
websockets==15.0.1
python-dotenv==1.1.0
pika==1.3.2
flake8==7.2.0